rename defaults to cli_defaults
