+++ components/sup/src/manager/mod.rs
@@ -47,8 +47,6 @@ use butterfly;
 use butterfly::member::Member;
 use butterfly::server::{timing::Timing, ServerProxy, Suitability};
 use butterfly::trace::Trace;
-use common::command::package::install::InstallSource;
-use common::ui::UIWriter;
 use common::ListenCtlAddr;
 use futures::prelude::*;
 use futures::sync::mpsc;
